src/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pytest
requests
sqlalchemy
passlib[bcrypt]
pydantic[email]
python-dotenv
pandas 